Switch to dhcpd.conf.d
This commit adds dhcpd.conf.d as a directory for containing
configuration file fragments.  This permits the dhcpd instance to be
configured almost entirely by the sg-dhcp role, but for additional
assets such as a file containing MAC addresses to be 'dropped in'.  A
script is also provided in /usr/local/sbin/dhcpd-reconf which will
reconfigure the dhcpd server and restart it.